+++ src/inspector/inspectoritempicker.cpp
@@ -3,12 +3,12 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "inspectoritempicker.h"
-#include "inspectorwebsocketconnection.h"
+#include "inspectorhandler.h"
 
 InspectorItemPicker::InspectorItemPicker(QObject* parent)
     : ItemPicker(parent) {}
 
 bool InspectorItemPicker::itemPicked(const QStringList& list) {
-  InspectorWebSocketConnection::itemsPicked(list);
+  InspectorHandler::itemsPicked(list);
   return true;
 }
