+++ src/src.pro
@@ -108,6 +108,7 @@ SOURCES += \
         hkdf.cpp \
         iaphandler.cpp \
         imageproviderfactory.cpp \
+        inspector/inspectorhandler.cpp \
         inspector/inspectoritempicker.cpp \
         inspector/inspectorutils.cpp \
         inspector/inspectorwebsocketconnection.cpp \
@@ -246,6 +247,7 @@ HEADERS += \
         hkdf.h \
         iaphandler.h \
         imageproviderfactory.h \
+        inspector/inspectorhandler.h \
         inspector/inspectoritempicker.h \
         inspector/inspectorutils.h \
         inspector/inspectorwebsocketconnection.h \
@@ -955,6 +957,7 @@ else:wasm {
             platforms/wasm/wasmnetworkrequest.cpp \
             platforms/wasm/wasmnetworkwatcher.cpp \
             platforms/wasm/wasmwindowcontroller.cpp \
+            platforms/wasm/wasminspector.cpp \
             systemtraynotificationhandler.cpp
 
     HEADERS += \
@@ -963,6 +966,7 @@ else:wasm {
             platforms/wasm/wasmauthenticationlistener.h \
             platforms/wasm/wasmnetworkwatcher.h \
             platforms/wasm/wasmwindowcontroller.h \
+            platforms/wasm/wasminspector.h \
             systemtraynotificationhandler.h
 
     SOURCES -= networkrequest.cpp
