+++ src/urlopener.cpp
@@ -5,7 +5,7 @@
 #include "urlopener.h"
 #include "constants.h"
 #include "logger.h"
-#include "inspector/inspectorwebsocketconnection.h"
+#include "inspector/inspectorhandler.h"
 #include "settingsholder.h"
 
 #include <QDesktopServices>
@@ -28,9 +28,9 @@ void UrlOpener::open(QUrl url, bool addEmailAddress) {
   }
 
   if (!Constants::inProduction()) {
-    InspectorWebSocketConnection::setLastUrl(url.toString());
+    InspectorHandler::setLastUrl(url.toString());
 
-    if (InspectorWebSocketConnection::stealUrls()) {
+    if (InspectorHandler::stealUrls()) {
       return;
     }
   }
