+++ tests/auth/auth.pro
@@ -46,7 +46,7 @@ HEADERS += \
     ../../src/featurelist.h \
     ../../src/hawkauth.h \
     ../../src/hkdf.h \
-    ../../src/inspector/inspectorwebsocketconnection.h \
+    ../../src/inspector/inspectorhandler.h \
     ../../src/ipaddress.h \
     ../../src/leakdetector.h \
     ../../src/logger.h \
@@ -74,7 +74,7 @@ HEADERS += \
 
 SOURCES += \
     mocmozillavpn.cpp \
-    ../unit/mocinspectorwebsocketconnection.cpp \
+    ../unit/mocinspectorhandler.cpp \
     ../../src/authenticationinapp/authenticationinapp.cpp \
     ../../src/authenticationinapp/authenticationinapplistener.cpp \
     ../../src/authenticationinapp/incrementaldecoder.cpp \
